+++ test/formatters/installers/pyang.bash
@@ -1,0 +1,2 @@
+apt-get install -y python3-pip
+pip install pyang