+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog].
 
 ## Unreleased
 ### Enhancements
+### Formatters
+### Bugs fixed
+* `apheleia-indent-lisp-buffer` updated to apply local variables after
+  calling major-mode. Also includes setting for `indent-tabs-mode` ([#286]).
+* [Formatter scripts](scripts/formatters) will now work on Windows if Emacs
+  can find the executable defined in the shebang.
+
+[#286]: https://github.com/radian-software/apheleia/pull/286
+[#285]: https://github.com/radian-software/apheleia/issues/285
+
+## 4.1 (released 2024-02-25)
+### Enhancements
 * Use `perltidy` as default formatter for `cperl-mode` ([#260]).
 * The `perltidy` formatter now supports Emacs indentation and line length
   settings ([#261]).
@@ -35,7 +47,10 @@ The format is based on [Keep a Changelog].
   clojurescript, edn files. ([#271])
 * Stylua is used now in `lua-ts-mode` as well as just `lua-mode`, by
   default ([#275]).
-* [`treefmt`](https://numtide.github.io/treefmt) for project configured formatters ([#280]).
+* [`treefmt`](https://numtide.github.io/treefmt) for project
+  configured formatters ([#280]).
+* [`pyang`](https://github.com/mbj4668/pyang) for
+  [YANG](https://www.rfc-editor.org/rfc/rfc7950.html) ([#282]).
 
 ### Bugs fixed
 * Apheleia sometimes failed to determine indent level from Emacs
@@ -55,6 +70,7 @@ The format is based on [Keep a Changelog].
 [#274]: https://github.com/radian-software/apheleia/issues/274
 [#275]: https://github.com/radian-software/apheleia/pull/275
 [#279]: https://github.com/radian-software/apheleia/pull/279
+[#282]: https://github.com/radian-software/apheleia/pull/282
 [#284]: https://github.com/radian-software/apheleia/pull/284
 
 ## 4.0 (released 2023-11-23)